Renamed vector to module and added reimports.
